feat(frontend): Migrate frontend to TypeScript
BREAKING CHANGE:
