+++ frontend/app/layout.tsx
@@ -1,0 +1,28 @@
+import './globals.css'
+
+import SessionProviderWrapper from '@/wrappers/SessionProviderWrapper'
+import type { Metadata } from 'next'
+import { Inter } from 'next/font/google'
+
+const inter = Inter({ subsets: ['latin'] })
+
+export const metadata: Metadata = {
+  title: 'PeerPrep',
+  description: 'Generated by create next app',
+}
+
+export default function RootLayout({
+  children,
+}: {
+  children: React.ReactNode
+}) {
+  return (
+    <html lang="en">
+      <body className={inter.className}>
+        <SessionProviderWrapper>
+          {children}
+        </SessionProviderWrapper>
+      </body>
+    </html>
+  )
+}
