+++ frontend/app/sign_up/page.js
@@ -1,7 +1,7 @@
 "use client";
 
-import Link from 'next/link';
 import axios from 'axios';
+import Link from 'next/link';
 import { useRouter } from 'next/navigation';
 import { useState } from 'react';
