+++ frontend/package.json
@@ -1,10 +1,6 @@
 {
   "name": "frontend",
   "version": "1.0.0",
-  "description": "this project contains the frontend code",
-  "author": "",
-  "license": "ISC",
-  "main": "index.js",
   "scripts": {
     "test": "jest",
     "dev": "next dev",
@@ -12,24 +8,33 @@
     "start": "next start",
     "lint": "next lint"
   },
-  "devDependencies": {
-    "jest": "^29.6.4",
-    "serve": "^14.2.1"
-  },
   "dependencies": {
-    "autoprefixer": "10.4.15",
-    "axios": "^1.5.0",
-    "clsx": "^2.0.0",
-    "eslint": "8.49.0",
-    "eslint-config-next": "13.4.19",
-    "next": "13.4.19",
-    "next-auth": "^4.23.1",
-    "nodemon": "^3.0.1",
-    "postcss": "8.4.29",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "react-hook-form": "^7.46.1",
-    "react-icons": "^4.11.0",
-    "tailwindcss": "3.3.3"
+    "autoprefixer": "latest",
+    "axios": "latest",
+    "clsx": "latest",
+    "eslint": "latest",
+    "eslint-config-next": "latest",
+    "next": "latest",
+    "next-auth": "latest",
+    "nodemon": "latest",
+    "postcss": "latest",
+    "react": "latest",
+    "react-dom": "latest",
+    "react-hook-form": "latest",
+    "react-icons": "latest",
+    "tailwindcss": "latest"
+  },
+  "devDependencies": {
+    "@types/node": "latest",
+    "@types/react": "latest",
+    "@types/react-dom": "latest",
+    "autoprefixer": "latest",
+    "eslint": "latest",
+    "eslint-config-next": "latest",
+    "jest": "latest",
+    "postcss": "latest",
+    "serve": "latest",
+    "tailwindcss": "latest",
+    "typescript": "latest"
   }
 }
