--- frontend/src/app/layout.js
@@ -1,25 +1,0 @@
-"use client";
-
-import '../styles/globals.css';
-import "../styles/text-editor.css";
-
-import { Inter } from 'next/font/google';
-
-const inter = Inter({ subsets: ['latin'] })
-
-export const metadata = {
-  title: 'PeerPrep',
-  description: 'Generated by create next app',
-}
-
-export default function RootLayout({
-  children
-}) {
-  return (
-    <html lang="en">
-      <body className={inter.className}>
-        {children}
-      </body>
-    </html>
-  )
-}
