--- frontend/src/app/page.js
@@ -1,134 +1,0 @@
-"use client";
-
-import { Button, LinkButton } from "../primitives/Button";
-
-import clsx from "clsx";
-import { getServerSession } from "next-auth/next";
-import { signIn } from "next-auth/react";
-import { SignInIcon } from "../icons";
-import { NavigationLayout } from "../layouts/Navigation";
-import { Container } from "../primitives/Container";
-import { authOptions } from "./api/auth/[...nextauth]";
-import styles from "./page.module.css";
-
-function Feature({
-  title,
-  description,
-  className,
-  ...props
-}) {
-  return (
-    <div className={clsx(className, styles.featuresFeature)} {...props}>
-      <h4 className={styles.featuresFeatureTitle}>{title}</h4>
-      <p className={styles.featuresFeatureDescription}>{description}</p>
-    </div>
-  );
-}
-
-export const metadata = {
-  title: "PeerPrep",
-  description: "Generated by create next app",
-};
-
-export default function Home() {
-  return (
-    <>
-      <NavigationLayout>
-        <Container className={styles.section}>
-          <div className={styles.heroInfo}>
-            <h1 className={styles.heroTitle}>
-              Crack the code with PeerPrep
-            </h1>
-            <p className={styles.heroLead}>
-              Join a community of driven learners and mentors to sharpen your
-              skills with PeerPrep.
-            </p>
-          </div>
-          <div className={styles.heroActions}>
-            <Button icon={<SignInIcon />} onClick={() => signIn()}>
-              Sign in
-            </Button>
-            <LinkButton
-              href="https://github.com/CS3219-AY2324S1/ay2324s1-course-assessment-g25"
-              target="_blank"
-              variant="secondary"
-            >
-              Learn more
-            </LinkButton>
-          </div>
-        </Container>
-        <Container className={styles.section}>
-          <h2 className={styles.sectionTitle}>Features</h2>
-          <div className={styles.featuresGrid}>
-            <Feature
-              description={
-                <>
-                  Feature 1
-                </>
-              }
-              title="Feature 1"
-            />
-            <Feature
-              description={
-                <>
-                  Feature 2
-                </>
-              }
-              title="Feature 2"
-            />
-            <Feature
-              description={
-                <>
-                  Feature 3
-                </>
-              }
-              title="Feature 3"
-            />
-            <Feature
-              description={
-                <>
-                  Feature 4
-                </>
-              }
-              title="Feature 4"
-            />
-            <Feature
-              description={
-                <>
-                  Feature 5
-                </>
-              }
-              title="Feature 5"
-            />
-            <Feature
-              description={
-                <>
-                  Feature 6
-                </>
-              }
-              title="Feature 6"
-            />
-          </div>
-        </Container>
-      </NavigationLayout>
-    </>
-  );
-}
-
-// If logged in, redirect to dashboard
-export const getServerSideProps = async ({ req, res }) => {
-  const session = await getServerSession(req, res, authOptions);
-
-  if (session) {
-    return {
-      redirect: {
-        permanent: false,
-        destination: "/home",
-      },
-    };
-  }
-
-  return {
-    props: {},
-  };
-};
