--- frontend/src/components/Logo/Logo.js
@@ -1,26 +1,0 @@
-import clsx from "clsx";
-import styles from "./Logo.modules.css";
-
-export function Logo({ 
-  className, 
-  ...props 
-}) {
-  return (
-    <div className={clsx(className, styles.logo)} {...props}>
-      <svg
-        className={styles.mark}
-        fill="none"
-        viewBox="0 0 32 32"
-        xmlns="http://www.w3.org/2000/svg"
-      >
-        <path
-          clipRule="evenodd"
-          d="M21.657 8H2l5.657 5.6v7.733L21.657 8ZM10.343 24H30l-5.657-5.6v-7.733L10.343 24Z"
-          fill="currentColor"
-          fillRule="evenodd"
-        />
-      </svg>
-      <span className={styles.wordmark}>PeerPrep</span>
-    </div>
-  );
-}
