--- frontend/src/components/Logo/index.js
@@ -1,2 +1,0 @@
-export * from "./Logo";
-
