--- frontend/src/components/Navigation/Header.module.css
@@ -1,10 +1,0 @@
-.header {
-  border-bottom: 1px solid var(--color-border);
-}
-
-.container {
-  display: flex;
-  height: var(--header-height);
-  justify-content: space-between;
-  place-items: center;
-}
