--- frontend/src/components/Navigation/index.js
@@ -1,3 +1,0 @@
-export * from "./Footer";
-export * from "./Header";
-
