--- frontend/src/icons/index.js
@@ -1,3 +1,0 @@
-export * from "./GitHub";
-export * from "./SignIn";
-
