--- frontend/src/layouts/Navigation/Navigation.modules.css
@@ -1,9 +1,0 @@
-.layout {
-  display: flex;
-  min-height: 100vh;
-  flex-direction: column;
-}
-
-.footer {
-  margin-top: auto;
-}
