--- frontend/src/layouts/Navigation/index.js
@@ -1,2 +1,0 @@
-export * from "./Navigation";
-
