--- frontend/src/primitives/Button/Button.js
@@ -1,59 +1,0 @@
-import clsx from "clsx";
-import Link from "next/link";
-import { forwardRef } from "react";
-import styles from "./Button.module.css";
-
-export const Button = forwardRef(({
-  variant = "primary",
-  icon, 
-  children,
-  className,
-  ...props
-}, ref) => (
-  <button
-    ref={ref}
-    className={clsx(
-      className,
-      styles.button,
-      icon && !children && styles.iconButton,
-      {
-        [styles.buttonPrimary]: variant === "primary",
-        [styles.buttonSecondary]: variant === "secondary",
-        [styles.buttonSubtle]: variant === "subtle",
-        [styles.buttonDestructive]: variant === "destructive",
-      }
-    )}
-    {...props}
-  >
-    {icon && <span className={styles.icon}>{icon}</span>}
-    {children && <span className={styles.label}>{children}</span>}
-  </button>
-));
-
-export function LinkButton({
-  variant = "primary",
-  icon,
-  children,
-  className,
-  ...props
-}) {
-  return (
-    <Link
-      className={clsx(
-        className,
-        styles.button,
-        icon && !children && styles.iconButton,
-        {
-          [styles.buttonPrimary]: variant === "primary",
-          [styles.buttonSecondary]: variant === "secondary",
-          [styles.buttonSubtle]: variant === "subtle",
-          [styles.buttonDestructive]: variant === "destructive",
-        }
-      )}
-      {...props}
-    >
-      {icon && <span className={styles.icon}>{icon}</span>}
-      {children && <span className={styles.label}>{children}</span>}
-    </Link>
-  );
-}
