--- frontend/src/primitives/Button/index.js
@@ -1,2 +1,0 @@
-export * from "./Button";
-
