--- frontend/src/primitives/Container/index.js
@@ -1,2 +1,0 @@
-export * from "./Container";
-
